Enemy State Machine and Level 1
Created a polished well working state machine for the enemies to use. It works completly with the basic enemy "Charred Walker".
Also created the framework for level 1. Most is finished but will def. be worked on more with more additions and polishing.

Next commit should have the full level finished and at least another enemy done if not two.