Cutscenes using Timeline, Level Polishing and Enemy Spawning
Implemented cutscenes using cinemachine and timeline and saved the first cutscene as a prefab. Polished up the level and completly rebaked the scene with new objects baked while the majority of smaller objects are not being baked, but instead illuminated by realtime light reflection probes.
Created some spawn in prefabs for the enemies as well.

Overall a lot of polishing to get things to feel smoother and the addition of some audio as well.
(Added a audio mixer as well to not only make everything sound better, but make it easer to adjust different volumes level in fromt he UI settings menu)